feat: refactor folder strcuture
