+++ src/editors/containers/ProblemEditor/components/EditProblemView/SettingsWidget/index.jsx
@@ -13,8 +13,8 @@ import ResetCard from './settingsComponents/ResetCard';
 import MatlabCard from './settingsComponents/MatlabCard';
 import TimerCard from './settingsComponents/TimerCard';
 import TypeCard from './settingsComponents/TypeCard';
-import GeneralFeedbackCard from './settingsComponents/GeneralFeedbackCard';
-import GroupFeedbackCard from './settingsComponents/GroupFeedbackCard';
+import GeneralFeedbackCard from './settingsComponents/GeneralFeedback/index';
+import GroupFeedbackCard from './settingsComponents/GroupFeedback/index';
 import SwitchToAdvancedEditorCard from './settingsComponents/SwitchToAdvancedEditorCard';
 import messages from './messages';
 import { showAdvancedSettingsCards } from './hooks';
