+++ src/editors/containers/ProblemEditor/components/EditProblemView/SettingsWidget/messages.js
@@ -194,41 +194,5 @@ export const messages = {
     defaultMessage: 'Switch To Advanced Editor',
     description: 'message to confirm that a user wants to use the advanced editor',
   },
-  noGroupFeedbackSummary: {
-    id: 'authoring.problemeditor.settings.GroupFeedbackSummary.message',
-    defaultMessage: 'None',
-    description: 'message to confirm that a user wants to use the advanced editor',
-  },
-  groupFeedbackSummary: {
-    id: 'authoring.problemeditor.settings.GroupFeedbackSummary.message',
-    defaultMessage: '{groupFeedback} {count, plural, =0 {} other {(+# more)}}',
-    description: 'summary of current feedbacks provided for multiple problems',
-  },
-  addGroupFeedbackButtonText: {
-    id: 'authoring.problemeditor.settings.addGroupFeedbackButtonText',
-    defaultMessage: 'Add group feedback',
-    description: 'addGroupFeedbackButtonText',
-  },
-  groupFeedbackInputLabel: {
-    id: 'authoring.problemeditor.settings.GroupFeedbackInputLabel',
-    defaultMessage: 'Group feedback will appear when a student selects a specific set of answers.',
-    description: 'label for group feedback input',
-  },
-  groupFeedbackSettingTitle: {
-    id: 'authoring.problemeditor.settings.GroupFeedbackSettingTitle',
-    defaultMessage: 'Group Feedback',
-    description: 'label for group feedback setting',
-  },
-  generalFeebackSettingTitle: {
-    id: 'authoring.problemeditor.settings.generalFeebackSettingTitle',
-    defaultMessage: 'General Feedback',
-    description: 'label for general feedback setting',
-  },
-  generalFeedbackInputLabel: {
-    id: 'authoring.problemeditor.settings.generalFeedbackInputLabel',
-    defaultMessage: 'Enter the feedback to appear when a student submits a wrong answer. This will be overrdidden if you add answer-specific feedback.',
-    description: 'label for general feedback input describing rules',
-  },
-
 };
 export default messages;
