+++ src/editors/containers/ProblemEditor/components/EditProblemView/SettingsWidget/settingsComponents/GeneralFeedback/hooks.js
@@ -1,0 +1,29 @@
+import { useState, useEffect } from 'react';
+import _ from 'lodash-es';
+import messages from './messages';
+import * as module from './hooks';
+
+export const state = {
+  summary: (val) => useState(val),
+};
+
+export const generalFeedbackHooks = (generalFeedback, updateSettings) => {
+  const [summary, setSummary] = module.state.summary({ message: '', values: {}, intl: false });
+
+  useEffect(() => {
+    if (_.isEmpty(generalFeedback)) {
+      setSummary({ message: messages.noGeneralFeedbackSummary, values: {}, intl: true });
+    } else {
+      setSummary({ message: generalFeedback.substring(0, 25), values: {}, intl: false });
+    }
+  }, [generalFeedback]);
+
+  const handleChange = (event) => {
+    updateSettings({ generalFeedback: event.target.value });
+  };
+
+  return {
+    summary,
+    handleChange,
+  };
+};
