+++ src/editors/containers/ProblemEditor/components/EditProblemView/SettingsWidget/settingsComponents/GeneralFeedback/messages.js
@@ -1,0 +1,18 @@
+export const messages = {
+  generalFeebackSettingTitle: {
+    id: 'authoring.problemeditor.settings.generalFeebackSettingTitle',
+    defaultMessage: 'General Feedback',
+    description: 'label for general feedback setting',
+  },
+  generalFeedbackInputLabel: {
+    id: 'authoring.problemeditor.settings.generalFeedbackInputLabel',
+    defaultMessage: 'Enter the feedback to appear when a student submits a wrong answer. This will be overrdidden if you add answer-specific feedback.',
+    description: 'label for general feedback input describing rules',
+  },
+  generalFeedbackDescription: {
+    id: 'authoring.problemeditor.settings.generalFeedbackInputDescription',
+    defaultMessage: 'Enter the feedback to appear when a student submits a wrong answer. This will be overrdidden if you add answer-specific feedback.',
+    description: 'description for general feedback input, clariying useage',
+  },
+};
+export default messages;
